backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.2.0
pillow>=10.2.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
# TensorFlow (optional — only needed to run training/train_disease.py and load disease_model.h5)
# tensorflow>=2.15.0
